i18n/l10n: fix HTML-escaping of translated strings (#3239)
With `newstyle=True`, the `safe` flag is added to all translated strings added with `{{ _("foo") }}` gettext call, hence no HTML-escaping happens. We do not make use of any format placeholders in our strings, hence `newstyle`, which makes such formatting easier, is irrelevant for us, and even adds the need to escape `%` with is otherwise interpreted as placeholder by Jinja2. A literal string injection but with assured HTML-escaping is reasonable in our case, which is what `newstyle=False` (default) provides.

Since with `newstyle=False`, Python format placeholders are not interpreted automatically, `%` does not need to be escaped, respectively it must not be escaped to appear as single character in the resulting HTML document. Only the source strings are updated here, while the translations will be bulk-updated via Weblate, to also trigger new MO files there.

The whole `jinja2.ext.i18n` extension in fact is unnecessary in our case. We don't use any of its features, but only need to pass the `gettext` function `_()` into the Jinja2 environment.

`trim_blocks=False` is the default and we do not have any XML templates, hence removing those.

Signed-off-by: MichaIng <micha@dietpi.com>